example/requirements.txt
Django==1.9.5
mysql-python
beautifulsoup4==4.4.0
django-admin-sortable2==0.6.3
django-allauth==0.23.0
django-angular==0.8.1
django-appconf==1.0.1
#django-classy-tags==0.7.2
#django-cms==3.2.3
django-compressor==1.6
#-e git+https://github.com/divio/django-filer.git@409f76ce285341e8f9f4adb3065ab8e011387de0#egg=django-filer-develop
django-filter==0.13.0
django-fsm==2.3.0
django-fsm-admin==1.2.2
django-haystack==2.4.0
django-ipware==1.1.1
#django-libsass==0.4
django-mptt==0.7.4
django-parler==1.6
django-polymorphic==0.8.1
#django-post-office==2.0.5
django-rest-auth==0.5.0
#django-sass-processor==0.3.4
django-sekizai==0.9.0
django-select2==5.5.0
#djangoshop-stripe==0.1.4
#django-treebeard==4.0
#djangocms-bootstrap3==0.1.1
#djangocms-cascade==0.8.1
#djangocms-text-ckeditor==2.8.1
djangorestframework==3.3.3
drf-haystack==1.5.6
easy-thumbnails==2.3
elasticsearch==1.7.0
html5lib==0.9999999
jsonfield==1.0.3
#libsass==0.8.3
#oauthlib==1.0.3
Pillow==2.9.0
#pluggy==0.3.1
#py==1.4.30
Pygments==2.0.2
#python-openid==2.2.5
pytz==2015.7
requests==2.7.0
#requests-oauthlib==0.5.0
six==1.9.0
#stripe==1.27.1
#Unidecode==0.4.18
urllib3==1.12
virtualenv==13.1.2
#wheel==0.24.0
git+ssh://git@bitbucket.org/excentrics/django-edw.git